blok support, oled on slave borked
