+++ keyboards/tweetydabird/lotus58/blok/config.h
@@ -1,0 +1,7 @@
+#define I2C_DRIVER I2CD0
+#define I2C0_SDA_PIN GP16
+#define I2C0_SCL_PIN GP17
+
+#define SERIAL_USE_MULTI_TRANSACTION
+#define SPLIT_MODS_ENABLE
+#define SPLIT_TRANSPORT_MIRROR
