+++ keyboards/tweetydabird/lotus58/blok/halconf.h
@@ -1,0 +1,3 @@
+#pragma once
+#include_next <halconf.h>
+#define HAL_USE_I2C TRUE
