+++ keyboards/tweetydabird/lotus58/blok/keymaps/default/rules.mk
@@ -1,2 +1,2 @@
 ENCODER_MAP_ENABLE = yes
-AUTO_SHIFT_ENABLE = yes
+CONVERT_TO = blok
