+++ keyboards/tweetydabird/lotus58/blok/mcuconf.h
@@ -1,0 +1,4 @@
+#pragma once
+#include_next <mcuconf.h>
+#undef RP_I2C_USE_I2C0
+#define RP_I2C_USE_I2C0 TRUE
