+++ keyboards/tweetydabird/lotus58/config.h
@@ -4,9 +4,9 @@
 #pragma once
 
 // Options not yet fully implemented in info.json
-#define RGBLIGHT_LIMIT_VAL 175
-#define RGBLIGHT_DEFAULT_VAL ( RGBLIGHT_LIMIT_VAL / 2 )
-#define RGBLIGHT_SLEEP
+//#define RGBLIGHT_LIMIT_VAL 175
+//#define RGBLIGHT_DEFAULT_VAL ( RGBLIGHT_LIMIT_VAL / 2 )
+//#define RGBLIGHT_SLEEP
 
 #define SPLIT_HAND_PIN B5
 #define SPLIT_LAYER_STATE_ENABLE
