--- keyboards/tweetydabird/lotus58/elite_c/rules.mk
@@ -1,1 +1,0 @@
-BOOTLOADER = atmel-dfu
