--- keyboards/tweetydabird/lotus58/nanoboot/rules.mk
@@ -1,2 +1,0 @@
-BOOTLOADER = qmk-hid
-BOOTLOADER_SIZE = 512
